requirements.txt
dvc[s3]==3.54.1
fastapi==0.135.1
httpx==0.28.1
joblib==1.4.2
numpy==1.26.4
pandas==2.0.1
Pillow==12.1.1
pydantic==2.12.5
pytest==8.3.2
python-multipart==0.0.22
scikit_learn==1.8.0
uvicorn==0.30.6